Ignore storage records for deleted volumes
Ceilometer currently collects storage amounts for volumes even after
they are deleted. There is no API to query for when a volume is deleted,
the only thing available is a query for if the volume exists at the
point of query.

We hack around this until upstream OpenStack is fixed by querying out
the underlying `deleted_at` information from the Cinder database,
storing it in a tab-separated `deleted-volumes.tsv` file in the
`logger-state` directory in a cron job.