Limit processing to 24h worth of data per run
The Ceilometer backend has a tendency to time out when asked to produce data for longer timespans. To prevent this from happening we limit the amount of backlog to process to at most 24 hours.

For a brand new deployment without any previous checkpoint date, we start at three weeks before the current time in order to cover the backlog, whose retention typically is one week on sites.