Skip and warn server instances which have no zone
Instance that have failed creation due to not being able to talk to block storage or for other reasons may be present in the data with no availability zone on record, which may cause upstream processing of records to fail.
The code now warns to alert the administrator about this and skips the record.